requirements-lambda.txt
# Sunday Album — Lambda container dependencies
# Excludes: rawpy (no DNG on web), torch/diffusers, matplotlib,
#           click, python-dotenv, pytest, mypy, ruff

# Core image processing
opencv-python-headless>=4.10.0
numpy>=1.26.4
scipy>=1.13.1
scikit-image>=0.24.0
Pillow>=10.4.0

# HEIC support — pillow-heif ships bundled libheif wheels (no system package needed)
pillow-heif>=0.18.0

# AI APIs
anthropic>=0.49.0
openai>=1.47.1

# AWS SDK (pre-installed in Lambda runtime but pin for container builds)
boto3>=1.35.0
botocore>=1.35.0
